refactor(doordash): apply the order splitter to the distilled payload
