+++ src/server/handlers/mcp-handler.ts
@@ -6,6 +6,7 @@ import { analytics } from '../services/analytics-service.js';
 import { finalizeSignin } from '../services/mcp-service.js';
 import { settings } from '../config.js';
 import { getAppHost } from '../utils/index.js';
+import { postprocessDistilled } from '../services/retailers/postprocess.js';
 import { ServerLogger as Logger } from '../utils/logger/index.js';
 
 interface BrandTool {
@@ -62,64 +63,6 @@ type PurchaseHistoryDetailsResponse = {
   content: Array<unknown> | Record<string, unknown>;
 };
 
-/**
- * Split DoorDash orders when store is an array (new format)
- * If store is not an array, pass through unchanged for backward compatibility
- */
-function splitDoorDashOrders(orders: Array<unknown>): Array<unknown> {
-  const processedOrders: Array<unknown> = [];
-
-  for (const order of orders) {
-    if (typeof order !== 'object' || order === null || !('store' in order)) {
-      processedOrders.push(order);
-      continue;
-    }
-
-    const orderObj = order as Record<string, unknown>;
-
-    // Only process if store is an array (new format)
-    if (!Array.isArray(orderObj.store)) {
-      processedOrders.push(order); // Pass through unchanged
-      continue;
-    }
-
-    const store = orderObj.store as Array<unknown>;
-    const summary = (orderObj.summary as Array<unknown>) || [];
-    const items = (orderObj.items as Array<unknown>) || [];
-    const storeUrl = (orderObj.store_url as Array<unknown>) || [];
-
-    // Find max length for splitting
-    const maxLength = Math.max(
-      store.length,
-      summary.length,
-      items.length,
-      storeUrl.length
-    );
-
-    // Extract non-array fields once, reuse for each split order
-    const baseOrder: Record<string, unknown> = {};
-    const splitKeys = new Set(['store', 'summary', 'items', 'store_url']);
-    Object.keys(orderObj).forEach((key) => {
-      if (!splitKeys.has(key)) {
-        baseOrder[key] = orderObj[key];
-      }
-    });
-
-    // Split into individual orders
-    for (let i = 0; i < maxLength; i++) {
-      const splitOrder: Record<string, unknown> = { ...baseOrder };
-      // Set array fields at index i (return as strings, not arrays)
-      splitOrder.store = store[i] ?? store[0] ?? '';
-      splitOrder.summary = summary[i] ?? summary[0] ?? '';
-      splitOrder.items = items[i] ?? items[0] ?? '';
-      splitOrder.store_url = storeUrl[i] ?? storeUrl[0] ?? '';
-      processedOrders.push(splitOrder);
-    }
-  }
-
-  return processedOrders;
-}
-
 export const handlePurchaseHistory = async (req: Request, res: Response) => {
   const { brandId } = req.params;
 
@@ -159,8 +102,8 @@ export const handlePurchaseHistory = async (req: Request, res: Response) => {
       typeof rawContent === 'string' ? JSON.parse(rawContent) : (rawContent as unknown[]);
   }
 
-  if (brandId === 'doordash' && Array.isArray(content)) {
-    content = splitDoorDashOrders(content);
+  if (Array.isArray(content)) {
+    content = postprocessDistilled(brandId, content);
   }
 
   if (content.length > 0) {
@@ -306,8 +249,8 @@ export const handleDpageUrl = async (req: Request, res: Response) => {
     content =
       typeof rawContent === 'string' ? JSON.parse(rawContent as string) : (rawContent as unknown[]);
   }
-  if (brandId === 'doordash' && Array.isArray(content)) {
-    content = splitDoorDashOrders(content);
+  if (Array.isArray(content)) {
+    content = postprocessDistilled(brandId, content);
   }
 
   res.json({ link_id: '', hosted_link_url: '', content });
@@ -372,8 +315,8 @@ export const handleDpageSigninCheck = async (req: Request, res: Response) => {
           dataFetchOk = true;
         }
 
-        if (brandId === 'doordash' && Array.isArray(content)) {
-          content = splitDoorDashOrders(content);
+        if (Array.isArray(content)) {
+          content = postprocessDistilled(brandId, content);
         }
       }
     } catch (err) {
